Откат версии библиотеки ArduinoJson с последней 6.21.5 на предыдущую стабильную 6.18.5

+++ libraries/ArduinoJson/extras/tests/JsonArray/add.cpp
@@ -1,5 +1,5 @@
 // ArduinoJson - https://arduinojson.org
-// Copyright © 2014-2023, Benoit BLANCHON
+// Copyright Benoit Blanchon 2014-2021
 // MIT License
 
 #include <ArduinoJson.h>
@@ -40,7 +40,7 @@ TEST_CASE("JsonArray::add()") {
 
 #ifdef HAS_VARIABLE_LENGTH_ARRAY
   SECTION("vla") {
-    size_t i = 16;
+    int i = 16;
     char vla[i];
     strcpy(vla, "world");
