+++ libraries/ArduinoJson/extras/tests/JsonObject/createNestedObject.cpp
@@ -1,5 +1,5 @@
 // ArduinoJson - https://arduinojson.org
-// Copyright © 2014-2023, Benoit BLANCHON
+// Copyright Benoit Blanchon 2014-2021
 // MIT License
 
 #include <ArduinoJson.h>
@@ -15,7 +15,7 @@ TEST_CASE("JsonObject::createNestedObject()") {
 
 #ifdef HAS_VARIABLE_LENGTH_ARRAY
   SECTION("key is a VLA") {
-    size_t i = 16;
+    int i = 16;
     char vla[i];
     strcpy(vla, "hello");
